Widen template-haskell version constraint to allow version <2.22
This allows freer-simple to be used with GHC >= 9.4. Prior versions of
template-haskell are not compatible with those GHC versions. Newer
version of of template-haskell seem to be fully compatible with
freer-simple. That is, tests still pass and I haven't noticed any
issues in my own project (which use freer-simple).

Fixes #40